requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.6.0
Django==4.0.4
django-ckeditor==6.4.1
django-crispy-forms==1.14.0
django-environ==0.8.1
django-js-asset==2.0.0
fonttools==4.33.3
gunicorn==20.1.0
html5lib==1.1
Pillow==9.1.1
pycparser==2.21
pydyf==0.2.0
pyphen==0.12.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.1.0
zopfli==0.2.1
